+++ src/main/java/glsl/language/utility/GlslSyntaxHighlighter.java
@@ -51,7 +51,7 @@ public class GlslSyntaxHighlighter extends SyntaxHighlighterBase {
             return OPERATOR_KEYS;
         }
         if (PRIMITIVE_TYPES_KEYWORDS.contains(tokenType) || STORAGE_QUALIFIER_KEYWORDS.contains(tokenType)||
-            CONTROL_FLOW_KEYWORDS.contains(tokenType)) {
+            CONTROL_FLOW_KEYWORDS.contains(tokenType) || MACRO_KEYWORD.contains(tokenType)) {
             return KEY_KEYS;
         }
         if (tokenType.equals(GlslTypes.IDENTIFIER)) {